@@ -11,7 +11,7 @@ repositories {
 }
 
 android {
-    compileSdkVersion 29
+    compileSdkVersion 30
     defaultConfig {
         javaCompileOptions {
             annotationProcessorOptions {
