@@ -1,14 +1,15 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
-    ext.kotlin_version = '1.2.71'
+    ext.kotlin_version = '1.4.32'
     ext.room_version = '1.1.1'
     repositories {
         google()
-        jcenter()
+        mavenCentral()
+        maven { url 'https://maven.google.com' }
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.2.1'
-        classpath 'com.google.gms:google-services:4.0.1'
+        classpath 'com.android.tools.build:gradle:4.2.2'
+        classpath 'com.google.gms:google-services:4.3.10'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
     }
 }
@@ -17,7 +18,8 @@ allprojects {
     repositories {
         maven { url "https://jitpack.io" }
         google()
-        jcenter()
+        mavenCentral()
+        maven { url 'https://maven.google.com' }
 
         maven {
             credentials {
