@@ -2,8 +2,8 @@ package com.butterflymx.butterflymxapiclient.features.opendoor
 
 import android.app.Activity
 import android.os.Bundle
-import android.support.v4.content.ContextCompat
-import android.support.v7.widget.RecyclerView
+import androidx.core.content.ContextCompat
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.navigation.Navigation
