@@ -1,7 +1,7 @@
 package com.butterflymx.butterflymxapiclient.signinsignup
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
+import androidx.fragment.app.Fragment
 import android.view.KeyEvent
 import android.view.LayoutInflater
 import android.view.View
