@@ -1,7 +1,7 @@
 package com.butterflymx.butterflymxapiclient.utils.mvp
 
-import android.support.v4.app.Fragment
-import android.support.v7.app.AlertDialog
+import androidx.fragment.app.Fragment
+import androidx.appcompat.app.AlertDialog
 import android.widget.Toast
 import com.butterflymx.butterflymxapiclient.R
 
